pyproject.toml
[project]
name = "llm-ensemble"
version = "0.1.0"
description = "Achieve consensus across multiple LLMs through an AI judge coordinator"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Roberto Pagliari", email = "your-email@example.com"}
]
requires-python = ">=3.12"
keywords = ["llm", "ai", "consensus", "langchain", "langgraph", "agent", "multi-model"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "deepagents>=0.1.0",
    "langchain>=1.2.3",
    "langchain-anthropic>=1.3.1",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.3",
    "langchain-openai>=1.1.7",
    "langchain-xai>=1.2.1",
    "langgraph>=1.0.5",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]

[project.urls]
Homepage = "https://github.com/yourusername/llm-ensemble"
Repository = "https://github.com/yourusername/llm-ensemble"
Issues = "https://github.com/yourusername/llm-ensemble/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
